EE-1095: consolidate auction::{internal => detail}
